+++ backend/FileService/FileService.UseCases/FilesStorage/IChunkSizeCalculator.cs
@@ -5,5 +5,5 @@ namespace FileService.UseCases.FilesStorage;
 
 public interface IChunkSizeCalculator
 {
-    Result<(long ChunkSizeBytes, int TotalChunks), Error> Calculate(long fileSizeBytes);
+    Result<(int ChunkSizeBytes, int TotalChunks), Error> Calculate(long fileSizeBytes);
 }
